pyproject.toml
[project]
name = "license-plate-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "emnist>=0.0",
    "matplotlib>=3.10.8",
    "numpy>=2.4.4",
    "opencv-python>=4.13.0.92",
    "pillow>=12.1.1",
    "tensorflow>=2.21.0",
]
